feiaiagent/website_deployment/metagpt_questionnaire/requirements_simple.txt
# Simplified requirements for network issues
requests>=2.25.0
urllib3>=1.26.0
pandas>=1.3.0
numpy>=1.21.0
aiohttp>=3.7.0
python-dotenv>=0.19.0
colorlog>=6.0.0
pydantic>=1.8.0
pytest>=6.0.0
